src/activities-api/pyproject.toml
[tool.poetry]
name = "activities-api"
version = "0.1.0"
description = ""
authors = ["Oriol Fabregas <oriol@fabreteca.com>"]
packages = [{include = "activities-api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
flask = "^3.0.0"
opentelemetry-exporter-otlp = "^1.20.0"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-requests = "^0.43b0"
opentelemetry-instrumentation-flask = "^0.43b0"
pymongo = "^4.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
